[refactor] 클린코드를 위한 폴더 및 파일 구조 변경
