+++ src/utils/utils.js
@@ -1,0 +1,3 @@
+export function formatPrice(price) {
+  return price.toLocaleString('ko-KR');
+}
